+++ apps/cowswap-frontend/src/modules/hooksStore/hooks/useAddHook.ts
@@ -1,11 +1,12 @@
 import { useSetAtom } from 'jotai'
 import { useCallback } from 'react'
 
+import { CowHookDetails } from '@cowprotocol/hook-dapp-lib'
+
 import { v4 as uuidv4 } from 'uuid'
 
 import { setHooksAtom } from '../state/hookDetailsAtom'
-import { AddHook, CowHookDetailsSerialized, HookDapp } from '../types/hooks'
-import { appendDappIdToCallData } from '../utils'
+import { AddHook, HookDapp } from '../types/hooks'
 
 export function useAddHook(dapp: HookDapp, isPreHook: boolean): AddHook {
   const updateHooks = useSetAtom(setHooksAtom)
@@ -15,16 +16,13 @@ export function useAddHook(dapp: HookDapp, isPreHook: boolean): AddHook {
       console.log('[hooks] Add ' + (isPreHook ? 'pre-hook' : 'post-hook'), hookToAdd, isPreHook)
 
       const uuid = uuidv4()
-      const hookDetails: CowHookDetailsSerialized = {
-        hookDetails: {
-          ...hookToAdd,
-          uuid,
-          hook: {
-            ...hookToAdd.hook,
-            callData: appendDappIdToCallData(hookToAdd.hook.callData, dapp.id),
-          },
+      const hookDetails: CowHookDetails = {
+        ...hookToAdd,
+        uuid,
+        hook: {
+          ...hookToAdd.hook,
+          dappId: dapp.id,
         },
-        dappId: dapp.id,
       }
 
       updateHooks((hooks) => {