+++ apps/cowswap-frontend/src/modules/hooksStore/index.ts
@@ -1,3 +1,4 @@
 export { HooksStoreWidget } from './containers/HooksStoreWidget'
 export { useHooks } from './hooks/useHooks'
 export { usePostHooksRecipientOverride } from './hooks/usePostHooksRecipientOverride'
+export { useCustomHookDapps } from './hooks/useCustomHookDapps'