+++ apps/cowswap-frontend/src/modules/hooksStore/state/customHookDappsAtom.ts
@@ -39,7 +39,7 @@ export const customPostHookDappsAtom = atom((get) => {
   return Object.values(get(customHookDappsAtom).post) as HookDappIframe[]
 })
 
-export const addCustomHookDappAtom = atom(null, (get, set, isPreHook: boolean, dapp: HookDappIframe) => {
+export const upsertCustomHookDappAtom = atom(null, (get, set, isPreHook: boolean, dapp: HookDappIframe) => {
   const { chainId } = get(walletInfoAtom)
   const state = get(customHookDappsInner)
 
@@ -72,7 +72,7 @@ export const removeCustomHookDappAtom = atom(null, (get, set, dapp: HookDappIfra
 
   // Delete applied hooks along with the deleting hook-dapp
   set(setHooksAtom, (hooksState) => ({
-    preHooks: (hooksState.preHooks || []).filter((hook) => hook.dappId !== hookDappId),
-    postHooks: (hooksState.postHooks || []).filter((hook) => hook.dappId !== hookDappId),
+    preHooks: (hooksState.preHooks || []).filter((hookDetails) => hookDetails.hook.dappId !== hookDappId),
+    postHooks: (hooksState.postHooks || []).filter((hookDetails) => hookDetails.hook.dappId !== hookDappId),
   }))
 })