+++ apps/cowswap-frontend/src/modules/trade/pure/TradeConfirmation/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react'
+import { ReactElement, useEffect, useRef, useState } from 'react'
 
 import {
   BackButton,
@@ -16,6 +16,9 @@ import ms from 'ms.macro'
 import { upToMedium, useMediaQuery } from 'legacy/hooks/useMediaQuery'
 import { PriceImpact } from 'legacy/hooks/usePriceImpact'
 
+import type { AppDataInfo } from 'modules/appData'
+
+import { OrderHooksDetails } from 'common/containers/OrderHooksDetails'
 import { CurrencyAmountPreview, CurrencyPreviewInfo } from 'common/pure/CurrencyInputPanel'
 
 import { QuoteCountdown } from './CountDown'
@@ -34,6 +37,7 @@ export interface TradeConfirmationProps {
 
   account: string | undefined
   ensName: string | undefined
+  appData?: string | AppDataInfo
   inputCurrencyInfo: CurrencyPreviewInfo
   outputCurrencyInfo: CurrencyPreviewInfo
   isConfirmDisabled: boolean
@@ -43,7 +47,7 @@ export interface TradeConfirmationProps {
   isPriceStatic?: boolean
   recipient?: string | null
   buttonText?: React.ReactNode
-  children?: JSX.Element
+  children?: ReactElement | ((restContent: ReactElement) => ReactElement)
 }
 
 export function TradeConfirmation(props: TradeConfirmationProps) {
@@ -70,6 +74,7 @@ export function TradeConfirmation(props: TradeConfirmationProps) {
     children,
     recipient,
     isPriceStatic,
+    appData,
   } = frozenProps || props
 
   /**
@@ -120,6 +125,10 @@ export function TradeConfirmation(props: TradeConfirmationProps) {
     onConfirm()
   }
 
+  const hookDetailsElement = (
+    <>{appData && <OrderHooksDetails appData={appData}>{(hookChildren) => hookChildren}</OrderHooksDetails>}</>
+  )
+
   return (
     <styledEl.WidgetWrapper onKeyDown={(e) => e.key === 'Escape' && onDismiss()}>
       <styledEl.Header>
@@ -142,8 +151,15 @@ export function TradeConfirmation(props: TradeConfirmationProps) {
             priceImpactParams={priceImpact}
           />
         </styledEl.AmountsPreviewContainer>
-        {children}
-        {/*Banners*/}
+        {typeof children === 'function' ? (
+          children(hookDetailsElement)
+        ) : (
+          <>
+            {children}
+            {hookDetailsElement}
+          </>
+        )}
+
         {showRecipientWarning && <CustomRecipientWarningBanner orientation={BannerOrientation.Horizontal} />}
         {isPriceChanged && !isPriceStatic && <PriceUpdatedBanner onClick={resetPriceChanged} />}
         <ButtonPrimary onClick={handleConfirmClick} disabled={isButtonDisabled} buttonSize={ButtonSize.BIG}>