+++ apps/widget-configurator/src/app/configurator/index.tsx
@@ -23,7 +23,13 @@ import ListItemText from '@mui/material/ListItemText'
 import Typography from '@mui/material/Typography'
 import { useWeb3ModalAccount, useWeb3ModalTheme } from '@web3modal/ethers5/react'
 
-import { COW_LISTENERS, DEFAULT_PARTNER_FEE_RECIPIENT, DEFAULT_TOKEN_LISTS, IS_IFRAME, TRADE_MODES } from './consts'
+import {
+  COW_LISTENERS,
+  DEFAULT_PARTNER_FEE_RECIPIENT_PER_NETWORK,
+  DEFAULT_TOKEN_LISTS,
+  IS_IFRAME,
+  TRADE_MODES,
+} from './consts'
 import { CurrencyInputControl } from './controls/CurrencyInputControl'
 import { CurrentTradeTypeControl } from './controls/CurrentTradeTypeControl'
 import { CustomImagesControl } from './controls/CustomImagesControl'
@@ -151,7 +157,7 @@ export function Configurator({ title }: { title: string }) {
     customColors: colorPalette,
     defaultColors: defaultPalette,
     partnerFeeBps,
-    partnerFeeRecipient: DEFAULT_PARTNER_FEE_RECIPIENT,
+    partnerFeeRecipient: DEFAULT_PARTNER_FEE_RECIPIENT_PER_NETWORK,
     standaloneMode,
     disableToastMessages,
     disableProgressBar,