+++ libs/hook-dapp-lib/src/types.ts
@@ -8,6 +8,7 @@ export interface CowHook {
   target: string
   callData: string
   gasLimit: string
+  dappId: string
 }
 
 export interface HookDappConditions {
@@ -17,7 +18,7 @@ export interface HookDappConditions {
 }
 
 export interface CowHookCreation {
-  hook: CowHook
+  hook: Omit<CowHook, 'dappId'>
   recipientOverride?: string
 }
 
@@ -26,12 +27,17 @@ export interface TokenData {
 }
 
 export interface CowHookDetails extends CowHookCreation {
+  hook: CowHook
+  uuid: string
+}
+
+export interface CowHookToEdit extends CowHookCreation {
   uuid: string
 }
 
 export interface CoWHookDappActions {
   addHook(payload: CowHookCreation): void
-  editHook(payload: CowHookDetails): void
+  editHook(payload: CowHookToEdit): void
   setSellToken(token: TokenData): void
   setBuyToken(token: TokenData): void
 }